# frontend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Village API",
    "version": "0.1.0"
  },
  "paths": {
...
    "\/build\/get": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "Get Build",
        "description": "Get a build.",
        "operationId": "get_build",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            },
            "name": "build_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/BuildWithMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/build\/list": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "List Builds",
        "description": "List all builds.",
        "operationId": "list_builds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/build\/runs": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "Get Build Runs",
        "description": "Get all runs for a build.",
        "operationId": "get_build_runs",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            },
            "name": "build_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/build\/delete": {
      "delete": {
        "tags": [
          "builds"
        ],
        "summary": "Delete Build",
        "description": "Delete a build.",
        "operationId": "delete_build",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            },
            "name": "build_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Build"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/run\/list": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "List all runs.",
        "operationId": "list_runs",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response List Runs",
                  "type": "array",
                  "items": {
                    "$ref": "#\/components\/schemas\/RunWithScript"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/run\/get": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "description": "Get a run.",
        "operationId": "get_run",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            },
            "name": "run_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/RunWithScriptDetailed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/run\/delete": {
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete Run",
        "description": "Delete a run.",
        "operationId": "delete_run",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            },
            "name": "run_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}